lsi-axm55xx: boot: give the RAM region execute attribute
In the past, the RAM section attribute was set to non-execute,
and the whole RAM section belong to domain 0.

Because the domain 0 was set to manager, which means accesses
are not checked against the permission bits in the translation
tables, so the problem wasn't exposed.

Now, a new commit set the domain 0 to client, the non-execute
RAM section will produce problems, so change the section attribute
to execute.

Signed-off-by: Wang Hui <Hui.Wang@windriver.com>
Signed-off-by: Calin Copos <calin.copos@windriver.com>