exynos5250: sata: Fixup SATA support for 3.10+
This static clock sources have gone so we need to add clock sources in
the device tree so the drivers can pick it up, and add the necessary
bindings between the i2c bus and the sata-phy layer.

There also seems to a change where the clocks don't get prepared before
they get enabled (perhaps because the clocks were always declared
statically?), so prepare and enable these clocks at the same time in the
users.  I have know idea if this is correct, but it seems to work.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>