LSI AXM55xx: Add PMU support
This patch adds PMU support to the Axxia AXM55xx platform. Note that
on this platform, all PMU IRQ lines are OR'ed together into a single
IRQ, and therefore, this implementation uses a rotating IRQ affinity
scheme to deal with it.

Take note that CONFIG_CPU_HAS_PMU was removed

Signed-off-by: David Mercado <david.mercado@windriver.com>
Signed-off-by: Calin Copos <calin.copos@windriver.com>